nvim: Change `nvim-surround` keymaps
Trying to follow the way that vim objects and actions work. To delete
a surround, instead of `sd`, it's better to use `ds` because it's more
consistent with other vim keyamps where to delete something, first comes
the action (delete) `d` and then text object, in this case a surround
represented by `s`. Same with changing a surround, better to use `cs` to
be more consistent. To add surrounds, I've mapped it to `s` since
I use it all the time and I want it to be short and easy, so `s` and `ss`
for current line should do.

* Lazy load `hop.nvim` and `nvim-surround` on corresponding keymaps

* zk: Use default picker