git: Add script to show modified files
Was trying to put this into an alias but for some obscure
reason the files aren't displayed relative to current path
within current git repo, instead it shows them relative to root.

This command is useful to do e.g `vim $(git modified)`. But if
you're not at the root of the repo and the files aren't displayed
relative, it won't work.

I suspect is has to do with how the command substitution for
`$(git rev-parse --show-toplevel)` is expanded when defined as an alias
in git config file.

Idea stolen from <https://github.com/nvie/git-toolbelt#git-modified>, I
just implemet it in a simplier way.