zsh: add fzf-tab plugin
Probably the best zsh plugin I've found so far and of the
biggest productivity boost I've had in some time.

ALL THE TIME I want/need to search through the options of a command,
the built-in tab completion is serviceable but it's lacking. What I
would do is enter the built-in scrollback of my terminal (kitty)
and then search with /, however there are a few pitfalls, first, if
the command has a lot of options you'll need to press tab multiple
times in order to be able to search for all options with the scrollback,
the reason is that when you initially trigger the completion, only a subset
of options are shown, so if you press tab one time and enter the scrollback,
you will only be able to search through the currently displayed options. And
this can get super annoying when a command really has _a lot_ of options. Second,
the search is not fuzzy.

This plugin has probably decreased the time I spend searching for an option by 2x times.