feat(tmux): replace copycat for fingers plugin
With copycat you have to cycle on each possible match until you get to
the one you want and I never could remember what were the keymaps for
each thing on copycat, with hints you just press whatever the key
combination is and done. Similar to sneak.vim or leap.nvim, tmux-fingers
gives you the same, but instead of jumping to those places, it is for
copying/opening. Also there's a single keymap that hints everything
relevant, so I don't have to remember a bunch of keymaps.

Alacritty and kitty have builtin hints but each one of them implements
it in a different way with different keymaps, using tmux is terminal
agnostic.

* add some comments for installing and removing plugins