fix(nvim.fzf): keep same behaviour for displaying filepaths
Putting the file at the beginning  (before the path) changes the
behaviour between fzf in the commandline and fzf in neovim. The query is
going to match on whatever is being passed to fzf, so if file is first
`file path/to/` instead of `path/to/file` then fuzzy works differently
in those cases. This is annoying because using fzf in the commandline is not
exactly the same as using it neovim.