refactor(nvim): replace oil for mini.files
Might sound silly but the creation of `oil:/` by
oil was what pushed me to replace it. I don't understand
why it has to create that folder whenever you do some action with
oil. I just find it annoying even if it doesn't show up in git.
I use nvim a lot to rename and move files in my system, which meant
having a `oil:/` folder everywhere, just no. Mini.files also allows
to edit files as a buffer, which is all I wanted from oil.

And I don't know if I'm crazy or what but I'm pretty sure that didn't
happen before, I've look into the issues and discussions of oil but
literally I have not found anything, no people complaining about it, or
how to disable it (if it's possible at all). I didn't feel like going
through the commit history to see when it started to happen.