feat(nvim): add ef-themes colorscheme
I collection of colorschemes ported from emacs[1] which I really like,
I don't actually use emacs and have never seriously use it (but I tried)
but I come across that collection a while ago. I think it is a wonderful
collection of colour pallets and I even tried porting them to nvim. But it
was a really tedious work figuring it out what each color should map to
the nvim version, so I stop. But thankfully someone did it.

[1]: https://protesilaos.com/emacs/ef-themes