fix(tmux): restore tmux sessions from terminal
Use terminals builtin features to run a command when
openging, instead of restoring on every single shell.

The problem was that for example, in gnome file manager
(nautilus) when you right click and select `open in console`,
it wouldn't work because tmux would restore sessions, so you
don't actually end up in the directory you want.