refactor(nvim): go back to oil.nvim
After a while using the files module from mini.nvim, there's a few pain
points I have with it:

* You're not editing files exactly like if it was a buffer because `h`
  and `l` are map for navigating the filesystem, don't know how I didn't
  notice that this would be a mayor problem when renaming files, it gets
  really annoying because `h` and `l` don't work the same.

* The preview window is to small so it really isn't useful, this was
  something I used a lot in oil

Besides that, I found a fix for what made me change from oil, which was
the creation of a `oil://` folder everywhere, I did search for it at the
time but didn't find anything. Today I found this on reddit:

https://www.reddit.com/r/neovim/comments/1g6f5jy/oilnvim_creating_empty_oil_directories/

Turns out it was simply because of the autocommand that creates missing
folders when saving a file in a folder that doesn't exist yet.

After searching a bit on github I found this solution which I liked the most:

https://github.com/tsakirist/dotfiles/blob/7d3454a57679e5ba1c8ce4273bbed3eb737bb99c/nvim/lua/tt/autocommands.lua#L117-L142

Ref: 76580a9