fix(zsh): enable completion of fzf options
For some reason completion for fzf options never worked for me, it does
work for every other command but not for fzf, weird. I decided to look
into into it today and found https://github.com/junegunn/fzf/issues/3349

* refactor(zsh.prompt): disable exit code in prompt. More problematic than I thought
with wrapping into next (ref: c86a488e). And besides, I already have it setup so
that the prompt symbol `❯` changes color to red when the exit code is other than 0.