fix(ghostty): use restore-tmux script inline
If I just use the executable from my scripts, there can be cases where
your shell env variables are not loaded correctly, because of some issue
with your config. This will make ghostty not find the executable because
the custom $PATH for scripts is not set, and you end up with an error when
opening ghostty and the shell doesn't load.

I ran into this a lot while figuring out d2c96774, so it's better to just
inline the entire script into the ghostty command option.