fix: add some files as templates
My mbsync systemd service for backing up emails has been throwing some
errors. It all came down to using ENV variables that were only set for
my user. Like for example trying to access executables in my
`~/.bin/scripts` folder like if they were available in the $PATH when
these services run, apparently they don't have access to my user ENV
variables. Even thou those services are only set locally for my user and
I have been using those scripts for months without issues until now.
I can also run the scripts those services run in without problems, so the
issues comes from the environment in which these services are executed when
they run under systemd.

At first I thought this was related to all the changes with d2c96774,
but after a bit of testing by putting the $PATH in `~/.zshenv` again,
and then running those services, it was still giving the same issue. The
only fix was giving the full path to an script (because it's not
available in the $PATH when it runs under systemd) and not assuming ENV
variables inside those scripts like $HOME for example. I have to
explicitly declare `HOME=/home/myuser` to make sure it all works.

Since I use chezmoi to manage my dotfiles. I've used the tmpl feature
to not hardcode those paths and instead use something like `.chezmoi.homeDir`
which is evaluated by chezmoi when you `apply` your dotfiles and results
in `/home/myuser`

Now, the weird thing is that I was running this setup for months without
issues, so either there was some update on systemd or something that no
longer allows services to access your user's ENV variables or indeed all
the changes with d2c96774 are the reason for this, on my quick testing
it wasn't that, but who knows. In any case, it's better to explicitly
declare variables in the scripts instead of relaying on variables that
may or may not be set externally.