refactor(zsh): don't save custom command history log
My idea was to have a custom command history log that also saved the
directory in which a command ran. And then make a script/program to
filter from it like getting the command history from a specific
directory. Filter by date and just grep/fzf through it like you would
normally do to the built-in history file.

But it never got anywhere, might revisit in the future because I still
think it's useful to search the command history I've run in a specific
folder. It's not very often but sometimes I do wish being able to search
history only from a specific folder.

There are also projects like atuin. I did try it and it's nice I guess,
but it does to much from what I need or want.

- https://github.com/atuinsh/atuin

* add comments explaining some of the prompt options