docs(aliasrc): add explanation about sudo alias
Was trying to get Ghostty shell integration working and added this:

- https://github.com/ghostty-org/ghostty/tree/b5912006ca0ac6b14815d9449ecd6cea422e28f2/src/shell-integration#zsh

```
if [[ -n $GHOSTTY_RESOURCES_DIR ]]; then
  source "$GHOSTTY_RESOURCES_DIR"/shell-integration/zsh/ghostty-integration
fi
```

But when doing so, I got this error:

```
/usr/share/ghostty/shell-integration/zsh/ghostty-integration:149: defining function based on alias `sudo'
/usr/share/ghostty/shell-integration/zsh/ghostty-integration:233: parse error near `()'
```

After a bit of research, it all came down to this alias `sudo='sudo '`.
I didn't have any idea why I was using that alias, and didn't document
it properly in my git history. But I was pretty sure that it was there
for a reason. AI made it pretty quick to figure out what it was for
instead of googling and having to go through a bunch of different websites.

I used perplexity. The reason for alias is to allow alias in expansion
for a command following sudo. After reading the response I remembered,
that, that was indeed the reason why I added the alias. I also found the
official documentation from gnu for bash:

- http://www.gnu.org/software/bash/manual/bashref.html#Aliases says: "If the
last character of the alias value is a space or tab character, then the next
command word following the alias is also checked for alias expansion."

* feat(ghostty): enable shell-integration for zsh

* refactor(tmux): remove laktak/extrakto since I never use it

= Reference

> Prompt:

Search the web for any shell/terminal/commandline issues where
one solution is setting alias sudo='sudo '

I found that in my aliases and I can't figure out why I added that
alias, and sadly I didn't document properly in git. But suspect it might
have been some issue that I was trying to resolve.

> Response:

The alias `alias sudo='sudo '` (with a trailing space) is used to enable
sudo to expand aliases for the command following sudo. Normally, aliases
in shells like bash or zsh are expanded only for the first word of
a command line, so when you prefix a command with sudo, the alias for
that command does not get expanded because sudo runs it in a clean
environment without alias expansion.

By defining `alias sudo='sudo '` (note the space at the end), you
effectively allow sudo to treat the next word as a separate command that
can be an alias itself. This works because in many shells, an alias that
ends with a space causes the shell to check the next word for an alias
as well. So, when you write `sudo e foo` and `e` is an alias, it will
expand `e` even when prefixed with sudo, making the alias usable with
sudo.

This trick is common to fix the issue where alias commands won't run
with sudo, letting you keep using your common aliases with elevated
privileges without redefining sudo-prefixed versions of the aliases. It
is effective in bash, zsh, and sh.

To summarize:

- Normally, aliases do not expand after sudo.
- Adding a space at the end of the sudo alias (`alias sudo='sudo '`)
  makes the shell check the next word for alias expansion.
- This lets you use your aliases with sudo seamlessly.
- This is why you might have created that alias—to solve the problem of
  aliases not working with sudo commands.

References from community discussions on Stack Overflow, Reddit, and
Arch Linux forums confirm this behavior and usage.[1][5][6]

[1](https://stackoverflow.com/questions/37209913/how-does-alias-sudo-sudo-work)
[2](https://stackoverflow.com/questions/66014021/bash-alias-using-sudo-returning-command-not-found)
[3](https://github.com/kovidgoyal/kitty/issues/6754)
[4](https://forum.manjaro.org/t/how-do-i-get-sudo-to-find-my-aliases-when-using-zsh/150578)
[5](https://bbs.archlinux.org/viewtopic.php?id=94026)
[6](https://www.reddit.com/r/linux/comments/13s57s/make_your_bashrc_aliases_work_with_sudo/)
[7](https://www.reddit.com/r/linux4noobs/comments/1bjrmb9/is_it_stupid_to_alias_ssudo_cause_im_lazy/)
[8](https://bbs.archlinux.org/viewtopic.php?id=283283)