fix(nvim.options): do not insert first completion suggestion with <CR>
The first item in completion menu is always "pre-selected", which makes
it easy to insert with <C-y>. However, using <CR> also inserts first
suggestion, which is not in the documentation or a default keymap (see
`help i_<CR>`), so it must be hardcoded/implicit. This results an
annoying edge case where I type, then I want to begin new line (<CR>),
but if completion popup is still showing, it just accepts first
suggestion and doesn't begin new line.

But even worse, when in combination with windwp/nvim-ts-autotag,
a plugin to autocomplete tags, type `<tag>` and it automatically adds
the tag resulting in `<tag>|</tag>` (cursor in | position). But the LSP
completion is still being triggered (possibly because the change was
made by a plugin and not detected properly by the LSP, because once the
cursor is in |, there should not be suggestions) so I still get
suggestions inside the tag, and pressing <CR> just accepts the first
completion suggestion, resulting in `<tag</tag>`. Yes, it just deletes
the closing `>`. This has been annoying me for months. Turns out that
it's fixed by simply adding `noselect` to completeopt. Which as the name
suggest, doesn't preselect anything.

The problem with using this, is that I was used to C-Y to insert first
suggestion, this can be just replace with C-N, but additionally, when
the first suggestion is a snippet. Just doing C-N doesn't trigger the
snippet, I have to do C-N+C-Y or C-N+Tab. But anyway, it's worth it,
that weird behaviour with tags has annoyed me for months, don't know why
I took so long to find the root of the issue. Additionally, using `noselect`
is even suggested in the official docs for lsp-completion, saying:

> Works best with completeopt=noselect.

See `help lsp-completion`